Codechange: Split baseset type definitions into separate files.
