+++ src/base_media_base.h
@@ -11,7 +11,6 @@
 #define BASE_MEDIA_BASE_H
 
 #include "fileio_func.h"
-#include "gfx_type.h"
 #include "textfile_type.h"
 #include "textfile_gui.h"
 #include "3rdparty/md5/md5.h"
@@ -221,119 +220,5 @@ template <class Tbase_set> /* static */ Tbase_set *BaseMedia<Tbase_set>::duplica
  */
 template <class Tbase_set>
 const char *TryGetBaseSetFile(const ContentInfo *ci, bool md5sum, const Tbase_set *s);
-
-/** Types of graphics in the base graphics set */
-enum GraphicsFileType : uint8_t {
-	GFT_BASE,     ///< Base sprites for all climates
-	GFT_LOGOS,    ///< Logos, landscape icons and original terrain generator sprites
-	GFT_ARCTIC,   ///< Landscape replacement sprites for arctic
-	GFT_TROPICAL, ///< Landscape replacement sprites for tropical
-	GFT_TOYLAND,  ///< Landscape replacement sprites for toyland
-	GFT_EXTRA,    ///< Extra sprites that were not part of the original sprites
-	MAX_GFT,      ///< We are looking for this amount of GRFs
-};
-
-/** Blitter type for base graphics sets. */
-enum BlitterType : uint8_t {
-	BLT_8BPP,       ///< Base set has 8 bpp sprites only.
-	BLT_32BPP,      ///< Base set has both 8 bpp and 32 bpp sprites.
-};
-
-struct GRFConfig;
-
-/** All data of a graphics set. */
-struct GraphicsSet : BaseSet<GraphicsSet, MAX_GFT, true> {
-private:
-	mutable std::unique_ptr<GRFConfig> extra_cfg; ///< Parameters for extra GRF
-public:
-	PaletteType palette;       ///< Palette of this graphics set
-	BlitterType blitter;       ///< Blitter of this graphics set
-
-	GraphicsSet();
-	~GraphicsSet();
-
-	bool FillSetDetails(const IniFile &ini, const std::string &path, const std::string &full_filename);
-	GRFConfig *GetExtraConfig() const { return this->extra_cfg.get(); }
-	GRFConfig &GetOrCreateExtraConfig() const;
-	bool IsConfigurable() const;
-	void CopyCompatibleConfig(const GraphicsSet &src);
-
-	static MD5File::ChecksumResult CheckMD5(const MD5File *file, Subdirectory subdir);
-};
-
-/** All data/functions related with replacing the base graphics. */
-class BaseGraphics : public BaseMedia<GraphicsSet> {
-public:
-	/** Values loaded from config file. */
-	struct Ini {
-		std::string name;
-		uint32_t shortname;                 ///< unique key for base set
-		uint32_t extra_version;             ///< version of the extra GRF
-		std::vector<uint32_t> extra_params; ///< parameters for the extra GRF
-	};
-	static inline Ini ini_data;
-
-};
-
-/** All data of a sounds set. */
-struct SoundsSet : BaseSet<SoundsSet, 1, true> {
-};
-
-/** All data/functions related with replacing the base sounds */
-class BaseSounds : public BaseMedia<SoundsSet> {
-public:
-	/** The set as saved in the config file. */
-	static inline std::string ini_set;
-
-};
-
-/** Maximum number of songs in the 'class' playlists. */
-static const uint NUM_SONGS_CLASS     = 10;
-/** Number of classes for songs */
-static const uint NUM_SONG_CLASSES    = 3;
-/** Maximum number of songs in the full playlist; theme song + the classes */
-static const uint NUM_SONGS_AVAILABLE = 1 + NUM_SONG_CLASSES * NUM_SONGS_CLASS;
-
-/** Maximum number of songs in the (custom) playlist */
-static const uint NUM_SONGS_PLAYLIST  = 32;
-
-/* Functions to read DOS music CAT files, similar to but not quite the same as sound effect CAT files */
-std::optional<std::string> GetMusicCatEntryName(const std::string &filename, size_t entrynum);
-std::optional<std::vector<uint8_t>> GetMusicCatEntryData(const std::string &filename, size_t entrynum);
-
-enum MusicTrackType : uint8_t {
-	MTT_STANDARDMIDI, ///< Standard MIDI file
-	MTT_MPSMIDI,      ///< MPS GM driver MIDI format (contained in a CAT file)
-};
-
-/** Metadata about a music track. */
-struct MusicSongInfo {
-	std::string songname;    ///< name of song displayed in UI
-	uint8_t tracknr;            ///< track number of song displayed in UI
-	std::string filename;    ///< file on disk containing song (when used in MusicSet class)
-	MusicTrackType filetype; ///< decoder required for song file
-	int cat_index;           ///< entry index in CAT file, for filetype==MTT_MPSMIDI
-	bool loop;               ///< song should play in a tight loop if possible, never ending
-	int override_start;      ///< MIDI ticks to skip over in beginning
-	int override_end;        ///< MIDI tick to end the song at (0 if no override)
-};
-
-/** All data of a music set. */
-struct MusicSet : BaseSet<MusicSet, NUM_SONGS_AVAILABLE, false> {
-	/** Data about individual songs in set. */
-	MusicSongInfo songinfo[NUM_SONGS_AVAILABLE];
-	/** Number of valid songs in set. */
-	uint8_t num_available;
-
-	bool FillSetDetails(const IniFile &ini, const std::string &path, const std::string &full_filename);
-};
-
-/** All data/functions related with replacing the base music */
-class BaseMusic : public BaseMedia<MusicSet> {
-public:
-	/** The set as saved in the config file. */
-	static inline std::string ini_set;
-
-};
 
 #endif /* BASE_MEDIA_BASE_H */
