+++ src/music.cpp
@@ -14,6 +14,7 @@
 /** The type of set we're replacing */
 #define SET_TYPE "music"
 #include "base_media_func.h"
+#include "base_media_music.h"
 #include "random_access_file_type.h"
 
 #include "safeguards.h"
