+++ src/network/core/tcp_content.cpp
@@ -13,6 +13,9 @@
 #include "../../textfile_gui.h"
 #include "../../newgrf_config.h"
 #include "../../base_media_base.h"
+#include "../../base_media_graphics.h"
+#include "../../base_media_music.h"
+#include "../../base_media_sounds.h"
 #include "../../ai/ai.hpp"
 #include "../../game/game.hpp"
 #include "../../fios.h"
