+++ src/network/network_content.cpp
@@ -15,6 +15,9 @@
 #include "../error.h"
 #include "../fileio_func.h"
 #include "../base_media_base.h"
+#include "../base_media_graphics.h"
+#include "../base_media_music.h"
+#include "../base_media_sounds.h"
 #include "../settings_type.h"
 #include "../strings_func.h"
 #include "network_content.h"
