+++ src/network/network_content_gui.cpp
@@ -15,6 +15,9 @@
 #include "../ai/ai.hpp"
 #include "../game/game.hpp"
 #include "../base_media_base.h"
+#include "../base_media_graphics.h"
+#include "../base_media_music.h"
+#include "../base_media_sounds.h"
 #include "../openttd.h"
 #include "../sortlist_type.h"
 #include "../stringfilter_type.h"
