+++ src/settings_gui.cpp
@@ -25,6 +25,9 @@
 #include "slider_func.h"
 #include "highscore.h"
 #include "base_media_base.h"
+#include "base_media_graphics.h"
+#include "base_media_music.h"
+#include "base_media_sounds.h"
 #include "company_base.h"
 #include "company_func.h"
 #include "viewport_func.h"
